+++ packages/flame_forge2d/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Next]
  - Added physics tied to widgets example
  - Added basic joint example
+ - Add generics passed to HasGameRef from PositionBodyComponent and SpriteBodyComponent
 
 ## [0.8.0-releasecandidate.13]
  - Update mechanism by which `BodyComponent`'s are disposed to use the `onRemove` method